Refactored replace_first in h2p.py as replace.py
